Fix highlighting
Newer versions of hugo support smartypants in markdown rendering, which
was changing the quotes to smart quotes. Luckily there's a shortcode
feature to disable this where necessary! Thumbs up for whoever was
thinking ahead.

Also, it seems we now have support for pygments CSS classes, so lanyon's
CSS works the way we originally intended!