Fix sidebar focus toggle
This diverges from upstream, but it fixes a visual regression that I
find too insidious to merge. When you close the sidebar, the toggle
remains colored, and you have to click elsewhere to take away focus and
un-color it. For more details, see #12. Thanks to @ryan-robeson who
investigated the issue.