Add a default .Site.Params.DateForm value
When lanyon-hugo is used as a theme (cloned into the themes/ directory)
rather than a standalone starter template, its config is ignored.

This meant that sites that didn't set DateForm in their config failed to
build. See #2 and #17.

Now, if DateForm isn't specified in the config, it will default to "Jan
2, 2006".